requirements.txt
anthropic>=0.40.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
